Move functions requiring wide dependencies to their own files.
This makes it easier to use gonum/plot without having to install
all of the dependencies if one does not care about them. For example,
if tiff, pdf, eps and x11 output are not required, it is easy to delete
the vg/vgimg/tiff.go file, vg/draw/formattedcanvas.go (depends on all
the formats known to plot), and the directories vg/vgpdf, vg/vgeps and
vg/vgx11.

I left pdf and jpg in vg/vgimg/vgimg.go, since they do not require a
third party library.
