+++ vg/draw/canvas.go
@@ -5,16 +5,11 @@
 package draw
 
 import (
-	"fmt"
 	"image/color"
 	"math"
 	"strings"
 
 	"github.com/gonum/plot/vg"
-	"github.com/gonum/plot/vg/vgeps"
-	"github.com/gonum/plot/vg/vgimg"
-	"github.com/gonum/plot/vg/vgpdf"
-	"github.com/gonum/plot/vg/vgsvg"
 )
 
 // A Canvas is a vector graphics canvas along with
@@ -225,39 +220,6 @@ func (CrossGlyph) DrawGlyph(c *Canvas, sty GlyphStyle, pt Point) {
 func New(c vg.CanvasSizer) Canvas {
 	w, h := c.Size()
 	return NewCanvas(c, w, h)
-}
-
-// NewFormattedCanvas creates a new vg.CanvasWriterTo with the specified
-// image format.
-//
-// Supported formats are:
-//
-//  eps, jpg|jpeg, pdf, png, svg, and tif|tiff.
-func NewFormattedCanvas(w, h vg.Length, format string) (vg.CanvasWriterTo, error) {
-	var c vg.CanvasWriterTo
-	switch format {
-	case "eps":
-		c = vgeps.New(w, h)
-
-	case "jpg", "jpeg":
-		c = vgimg.JpegCanvas{Canvas: vgimg.New(w, h)}
-
-	case "pdf":
-		c = vgpdf.New(w, h)
-
-	case "png":
-		c = vgimg.PngCanvas{Canvas: vgimg.New(w, h)}
-
-	case "svg":
-		c = vgsvg.New(w, h)
-
-	case "tif", "tiff":
-		c = vgimg.TiffCanvas{Canvas: vgimg.New(w, h)}
-
-	default:
-		return nil, fmt.Errorf("unsupported format: %q", format)
-	}
-	return c, nil
 }
 
 // NewCanvas returns a new (bounded) draw.Canvas of the given size.
